+++ src/resolvers/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { mergeResolvers } = require('graphql-tools')
+const { mergeResolvers } = require('@graphql-tools/merge')
 
 module.exports = mergeResolvers([
 	require('./tokens'),