+++ src/utils/config.js
@@ -2,7 +2,7 @@
 
 const { day } = require('./times')
 
-// Must be a function or object that load and returns the env variables at runtime.
+// Must be a function or object that loads and returns the env variables at runtime.
 // Otherwise it wouldn't be possible to mock the env variables with mockedEnv.
 module.exports = new Proxy({}, {
 	get: function(target, prop) {
@@ -11,6 +11,7 @@ module.exports = new Proxy({}, {
 			port: process.env.ACKEE_PORT || process.env.PORT || 3000,
 			dbUrl: process.env.ACKEE_MONGODB || process.env.MONGODB_URI,
 			allowOrigin: process.env.ACKEE_ALLOW_ORIGIN,
+			autoOrigin: process.env.ACKEE_AUTO_ORIGIN === 'true',
 			username: process.env.ACKEE_USERNAME,
 			password: process.env.ACKEE_PASSWORD,
 			isDemoMode: process.env.ACKEE_DEMO === 'true',