+++ src/utils/findMatchingOrigin.js
@@ -1,10 +1,21 @@
 'use strict'
 
-module.exports = (req, allowedOrigins) => {
+const fullyQualifiedDomainNames = require('./fullyQualifiedDomainNames')
+
+const findOrigin = (req, origins) => {
+	return origins.find((origin) => origin.includes(req.headers.origin) || origin.includes(req.headers.host))
+}
+
+module.exports = async (req, allowedOrigins, autoOrigin) => {
+	if (autoOrigin === true) {
+		const origins = await fullyQualifiedDomainNames()
+		return findOrigin(req, origins)
+	}
+
 	if (allowedOrigins === '*') return '*'
 
 	if (allowedOrigins != null) {
 		const origins = allowedOrigins.split(',')
-		return origins.find((origin) => origin.includes(req.headers.origin) || origin.includes(req.headers.host))
+		return findOrigin(req, origins)
 	}
 }